% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{add_commas}
\alias{add_commas}
\title{add_commas}
\usage{
add_commas(value)
}
\arguments{
\item{value}{Value to be re-formatted.}
}
\value{
Checked uploaded locations (added peril_id).
}
\description{
Inserts commas at thousands.
}
