% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelParams_funs.R
\name{advancedConfig_funs}
\alias{advancedConfig_funs}
\title{Advanced Output Configuration functions}
\usage{
advancedConfig_funs(session, model_settings, configurable)
}
\arguments{
\item{session}{Current session.}

\item{model_settings}{Model settings retrieved from the API.}

\item{configurable}{Whether model is configurable or not}
}
\value{
List of UI elements (input/selector widgets).
}
\description{
Functions for the Advanced Output Configuration/Rerun Model Parameters view.
}
