% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleAnaWorkflowSteps.R
\docType{data}
\name{analysisWorkflowSteps}
\alias{analysisWorkflowSteps}
\title{analysisWorkflowSteps}
\format{An object of class \code{list} of length 3.}
\usage{
analysisWorkflowSteps
}
\value{
List.
}
\description{
analysisWorkflowSteps
}
\keyword{datasets}
