% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFly_module.R
\name{buildFly}
\alias{buildFly}
\alias{buildFlyUI}
\title{Hazard Map Module}
\usage{
buildFlyUI(id)

buildFly(input, output, session, portfolioID, modelID, supplierID,
  versionID, analysisID, counter, active = reactive(TRUE))
}
\arguments{
\item{id}{Namespace ID for the module.}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{portfolioID}{Selected portfolio ID.}

\item{modelID}{Selected model ID.}

\item{supplierID}{Selected supplier model ID.}

\item{versionID}{Selected model ID version.}

\item{analysisID}{Selected analysis ID.}
}
\description{
Shiny Module for showing fly details.
}
\section{Functions}{
\itemize{
\item \code{buildFlyUI}: Returns the UI elements of the module.

\item \code{buildFly}: Allows user to build their own mode on the fly.
}}

