% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasisui_utils.R
\name{capitalize_names_df}
\alias{capitalize_names_df}
\title{Capitalize names of a dataframe}
\usage{
capitalize_names_df(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
df
}
\description{
Returns a dataframe with names with the first letterof all words capitalized.
}
