% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_loc.R
\name{check_loc}
\alias{check_loc}
\title{check_loc}
\usage{
check_loc(analysisID, portfolioID, data_hub)
}
\arguments{
\item{analysisID}{Selected analysis id.}

\item{portfolioID}{selected portfolio ID.}

\item{data_hub}{data hub stored in session$userData$data_hub}
}
\value{
Checked uploaded locations (added peril_id).
}
\description{
Compares uploaded and modeled locations.
}
