% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHazardMap_module.R
\name{createHazardMap}
\alias{createHazardMap}
\alias{createHazardMapUI}
\title{Hazard Map Module}
\usage{
createHazardMapUI(id)

createHazardMap(input, output, session, map_data, pins, analysisID)
}
\arguments{
\item{id}{Namespace ID for the module.}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{map_data}{Reactive expression yielding Leaflet map data.}

\item{pins}{Reactive expression yielding location pins as a \code{data.frame}
with variables \code{locnumber}, \code{longitude}, \code{latitude}
(case-insensitive).}

\item{analysisID}{Chosen analysis.}
}
\description{
Shiny Module for rendering a hazard map using Leaflet.
}
\section{Functions}{
\itemize{
\item \code{createHazardMapUI}: Returns the UI elements of the module.

\item \code{createHazardMap}: Defines the server logic of the module.
}}

\examples{
# TODO

}
