% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_config_module.R
\name{def_out_configUI}
\alias{def_out_configUI}
\alias{def_out_config}
\title{modeldetailsUI}
\usage{
def_out_configUI(id)

def_out_config(
  input,
  output,
  session,
  analysisID = reactive(NULL),
  analysisName = reactive(""),
  ana_flag = reactive("C"),
  counter = reactive(NULL),
  active = reactive(TRUE),
  customModSettings = reactive(NULL)
)
}
\arguments{
\item{analysisID}{Selected analysis ID.}

\item{analysisName}{Selected analysis name.}

\item{ana_flag}{Flag to know if the user is creating a new output configuration or rerunning an analysis.}

\item{counter}{Counter to decide whether to clear output options and show the output config panel}

\item{customModSettings}{Customizable Model settings.}
}
\value{
ana_flag flag to know if the user is creating a new output configuration or rerunning an analysis.

ana_post_status status of posting the analysis.
}
\description{
UI side of function wrapping panel to show oputput configuration.

Server side of function wrapping panel to show oputput configuration.
}
