% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_settings.R
\docType{data}
\name{default_tags}
\alias{default_tags}
\title{Default Tags}
\format{An object of class \code{character} of length 3.}
\usage{
default_tags
}
\description{
Set of default tags for output parameters
}
\keyword{datasets}
