% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputConfigUI_funs.R
\name{dynamicUI}
\alias{dynamicUI}
\title{Dynamic UI}
\usage{
dynamicUI(session, analysisID, tag, n_field, oed_field)
}
\arguments{
\item{session}{Current session.}

\item{analysisID}{Current analysis ID.}

\item{tag}{Current tag.}

\item{n_field}{Number of fields to produce.}

\item{oed_field}{Oed fields.}
}
\value{
UI element.
}
\description{
Add input fields for custom and drill-down configuration.
}
