% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputConfigUI_funs.R
\name{dynamicUI_btns}
\alias{dynamicUI_btns}
\title{Dynamic UI with buttons}
\usage{
dynamicUI_btns(
  session,
  analysisID,
  ana_flag,
  tag,
  oed_field,
  analysisSettings = NULL
)
}
\arguments{
\item{session}{Current session.}

\item{analysisID}{Current analysis ID.}

\item{ana_flag}{Analysis flag.}

\item{tag}{Current tag.}

\item{oed_field}{Oed fields.}

\item{analysisSettings}{Analysis settings for case where we re-run from settings template.}
}
\value{
UI element.
}
\description{
Add "+" and "x" buttons to dynamic UI.
}
