% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_settings.R
\docType{data}
\name{filesListDataNames}
\alias{filesListDataNames}
\title{filesListDataNames}
\format{
Named \code{list} of filesListData column names.
}
\usage{
filesListDataNames
}
\description{
Creating List for  col names of filesListData
}
\keyword{datasets}
