% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_config_settings.R
\docType{data}
\name{granToOed}
\alias{granToOed}
\title{granToOed}
\format{\code{data.frame} of granularities for output configuration.}
\usage{
granToOed
}
\description{
Data frame of variables, with respective oed field and info from analysis_settings.json.
}
\keyword{datasets}
