% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_text.R
\docType{data}
\name{helper-text}
\alias{helper-text}
\alias{file_Viewer}
\alias{landing_page}
\alias{defineSingleAna}
\alias{dashboard}
\title{Helper text}
\format{Each object is a named list of character strings, the names
  corresponding to the button IDs.}
\usage{
file_Viewer

landing_page

defineSingleAna

dashboard
}
\description{
Helper texts to show explanations next to buttons, one per file (?).
}
\keyword{datasets}
