% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_module.R
\name{linePlot}
\alias{linePlot}
\title{linePlot}
\usage{
linePlot(xlabel, ylabel, titleToUse, data)
}
\arguments{
\item{xlabel}{label X axis}

\item{ylabel}{label y axis}

\item{titleToUse}{title}

\item{data}{dataframe to plot}
}
\description{
line plot
}
\details{
linePlot structure
}
