% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{loginfo}
\alias{loginfo}
\alias{logMessage}
\title{loginfo}
\usage{
loginfo(..., logger)

logMessage(msg)
}
\arguments{
\item{...}{Extra arguments.}

\item{logger}{oasisui.module.}

\item{msg}{Character string message.}
}
\value{
Log in info.
}
\description{
loginfo
}
