% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_config_settings.R
\docType{data}
\name{LosstypesChoices}
\alias{LosstypesChoices}
\title{LosstypesChoices}
\format{Named \code{list} of Losstypes Choices. (Name empty)}
\usage{
LosstypesChoices
}
\description{
List of variables per granularity
}
\keyword{datasets}
