% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeldetails_module.R
\name{modeldetails}
\alias{modeldetails}
\alias{modeldetailsUI}
\title{Hazard Map Module}
\usage{
modeldetailsUI(id)

modeldetails(
  input,
  output,
  session,
  analysisID,
  modelID,
  portfolioID,
  file_pins,
  counter,
  active = reactive(TRUE)
)
}
\arguments{
\item{id}{Namespace ID for the module.}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{analysisID}{Selected analysis ID.}

\item{modelID}{Selected model ID.}

\item{portfolioID}{Selected portfolio ID.}

\item{file_pins}{file with coordiantes of exposure locations}

\item{counter}{Reactive value to trigger inputs download.}

\item{active}{reactive expression whether the module state should be updated.}
}
\description{
Shiny Module for showing model details.
}
\section{Functions}{
\itemize{
\item \code{modeldetailsUI}: Returns the UI elements of the module.

\item \code{modeldetails}: Defines the server logic of the module.
}}

