% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasisui-package.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Variants of standard operators}
\usage{
'See the help for match in base R.'
}
\arguments{
\item{...}{Arguments to \code{\link{\%in\%}}.}
}
\description{
Negation of \code{\link{\%in\%}} from R's \code{base} library.
}
