% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasisui_utils.R
\name{nothingToShowTable}
\alias{nothingToShowTable}
\title{Nothing to show in table}
\usage{
nothingToShowTable(contentMessage = "No data found")
}
\arguments{
\item{contentMessage}{Message to be displayed.}
}
\description{
Shows message in case table is table.
}
