% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasisapi.R
\docType{class}
\name{OasisAPI}
\alias{OasisAPI}
\title{OasisAPI}
\format{\code{\link{R6Class}} object.}
\usage{
OasisAPI
}
\value{
Object of \code{\link{R6Class}} with methods for connecting to Oasis API.
}
\description{
R6 Class for OasisUI API Hub.
}
\section{Arguments}{

\describe{
\item{user}{string for username.}
\item{pwd}{string for user password.}
}
}

\section{Methods}{

\describe{

Initialize
 \item{\code{api_init(host, port)}}{Initialize api and set url private.}
 \item{\code{get_url()}}{Return private url.}
 \item{\code{get_conn_init()}}{Return conn_init.}
 \item{\code{get_http_type()}}{Return http_type.}
 api response
 \item{\code{api_handle_response(response)}}{Handles api response.}
 \item{\code{api_fetch_response(meth, args, logMessage = print)}}{Fetches api response.}
 access token
 \item{\code{api_post_access_token(user, pwd)}}{Returns response of api access token.}
 \item{\code{set_tokens(user, pwd)}}{Set private access token and private refresh token.}
 \item{\code{get_access_token()}}{Return private access token.}
 refresh token
 \item{\code{get_refresh_token()}}{Return private refresh token.}
 \item{\code{api_post_refresh_token()}}{Post refresh token.}
 version
 \item{\code{set_version()}}{Set private version.}
 \item{\code{get_version()}}{Return private version.}
 healtcheck
 \item{\code{api_get_healthcheck()}}{Perform api healthcheck.}
 api query
 \item{\code{api_basic_query(query_path_basic, query_list, query_method, ...)}}{Construct query without version for the api.}
 \item{\code{api_query(query_path, query_list, query_method, ...)}}{Construct query with version for the api.}
 \item{\code{api_get_query(query_path, query_list, ...)}}{Construct GET query to the api.}
 \item{\code{api_post_query(query_path, query_list, ...)}}{Construct POST query to the api.}
 \item{\code{api_delete_query(query_path, query_list, ...)}}{Construct DELETE query to the api.}
 \item{\code{api_patch_query(query_path, query_list, ...)}}{Construct PATCH query to the api.}
 \item{\code{api_post_file_query(query_path, query_body = NULL, ...)}}{POST file query to the api.}
 \item{\code{api_body_query(query_path, query_body = NULL, query_method = "POST", ...)}}{PUT/POST query with body field to the api.}
 \item{\code{api_return_query_res(query_path, query_list = NULL, query_method, ...)}}{PUT/POST query with body field to the api.}
 return from query
 \item{\code{return_df(query_path, api_param = "", query_method = "GET"))}}{MAnipulate data returned from GET query.}
 write to disk
 \item{\code{api_get_analyses_tar(id, label, dest = tempfile(fileext = ".tar"))}}{GET files and write to disk tar bundle.}
}
}

\section{Usage}{

\preformatted{oasisapi <- OasisAPI$new()
oasisapi$method(arg)
}
}

\keyword{datasets}
