% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingo-package.R
\docType{package}
\name{oasisui-package}
\alias{oasisui}
\alias{oasisui-package}
\title{oasisui: Oasis User Interface}
\description{
This package provides an extensible implementation of a user
  interface for modelling workflows. The interface is a web application based
  on the popular Shiny framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oasislmf.github.io/}
  \item \url{https://github.com/OasisLMF/OasisUI}
  \item Report bugs at \url{https://github.com/OasisLMF/OasisUI/issues}
}

}
\author{
\strong{Maintainer}: Oasis LMF \email{support@oasislmf.org} [copyright holder]

Other contributors:
\itemize{
  \item Mirai Solutions GmbH \email{info@mirai-solutions.com} [contractor]
}

}
