% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_server.R
\docType{data}
\name{OASISUI_GUEST_ID}
\alias{OASISUI_GUEST_ID}
\title{OASISUI_GUEST_ID}
\format{
An object of class \code{character} of length 1.
}
\usage{
OASISUI_GUEST_ID
}
\value{
Set value.
}
\description{
Id to use for an unidentified user.
}
\keyword{datasets}
