% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{oasisuiButton}
\alias{oasisuiButton}
\title{oasisuiButton}
\usage{
oasisuiButton(
  inputId,
  label,
  icon = NULL,
  width = NULL,
  class = c("btn", "btn-primary"),
  ...
)
}
\arguments{
\item{inputId}{widget id}

\item{label}{widget label}

\item{icon}{widget icon}

\item{width}{widget width}

\item{class}{HTML class attribute.}

\item{...}{Arguments to \code{\link[shiny:actionButton]{shiny::actionButton()}}.}
}
\value{
Button widget (UI element).
}
\description{
Modified version of the default \code{\link[shiny:actionButton]{shiny::actionButton()}}.
}
