% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{oasisuiCheckboxButton}
\alias{oasisuiCheckboxButton}
\title{oasisuiCheckboxButton}
\usage{
oasisuiCheckboxButton(inputId, label, icon = NULL, width = NULL,
  class = c("btn"), ...)
}
\arguments{
\item{inputId}{widget id}

\item{label}{widget label}

\item{icon}{widget icon}

\item{width}{widget width}

\item{class}{HTML class attribute.}

\item{...}{Arguments to \code{\link[shiny:actionButton]{shiny::actionButton()}}.}
}
\value{
List of tags.
}
\description{
Modified version of the default \code{\link[shiny:actionButton]{shiny::actionButton()}}.
}
