% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasisuiPanel.R
\name{oasisuiPanelHeading}
\alias{oasisuiPanelHeading}
\title{Panel Heading}
\usage{
oasisuiPanelHeading(heading)
}
\arguments{
\item{heading}{The desired heading, either a character string (which gets
properly wrapped) or a tags list (returned as-is).}
}
\description{
Construct an appropriate heading for a \link{oasisuiPanel}.
}
