% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasisuiPanel.R
\name{oasisuiPanelHeadingOutput}
\alias{oasisuiPanelHeadingOutput}
\title{Panel Heading UI Output}
\usage{
oasisuiPanelHeadingOutput(outputId, ...)
}
\arguments{
\item{outputId}{ID of the dynamical panel heading.}

\item{...}{Other arguments to pass to the container tag function. This is
useful for providing additional classes for the tag.}
}
\description{
Panel heading UI placeholder to be rendered dynamically via
\code{\link[=renderOasisuiPanelHeading]{renderOasisuiPanelHeading()}}.
}
\seealso{
\code{\link[=renderOasisuiPanelHeading]{renderOasisuiPanelHeading()}}
}
