% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{oasisuiRefreshButton}
\alias{oasisuiRefreshButton}
\title{oasisuiRefreshButton}
\usage{
oasisuiRefreshButton(id, width = NULL)
}
\arguments{
\item{id}{ID of the button}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link{validateCssUnit}}.}
}
\description{
oasisuiRefreshButton
}
