% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasisuiTable_module.R
\name{oasisuiTableUI}
\alias{oasisuiTableUI}
\alias{oasisuiTable}
\title{oasisuiTableUI}
\usage{
oasisuiTableUI(id)

oasisuiTable(input, output, session, data, selection = "none",
  escape = TRUE, filter = FALSE, rownames = TRUE,
  preselRow = reactive({     NULL }), maxrowsperpage = 10)
}
\arguments{
\item{id}{Namespace ID for the module.}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{data}{dataframe to show in table.}

\item{selection}{param of datatable, default"none".}

\item{escape}{param of datatable, default TRUE.}

\item{filter}{param of datatable, default FALSE.}

\item{rownames}{param of datatable, default TRUE.}

\item{preselRow}{reactive of preselected row default reactive({NULL}).}

\item{maxrowsperpage}{param of datatable, default 10.}
}
\value{
List of tags.

rows_selected reactive of selected rows as returned from datatable.

rows_current reactive of current rows as returned from datatable.
}
\description{
Server logic to show the oasisui table output
}
