% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{oasisuiTableButton}
\alias{oasisuiTableButton}
\title{oasisuiTableButton}
\usage{
oasisuiTableButton(inputId, label, icon = NULL, width = NULL,
  class = c("btn", "btn-secondary"), ...)
}
\arguments{
\item{...}{Arguments to \code{\link[oasisui:oasisuiButton]{oasisui::oasisuiButton()}}.}
}
\value{
Acces to Oasis UI.
}
\description{
Modified version of the default \code{\link[oasisui:oasisuiButton]{oasisui::oasisuiButton()}}.
}
