% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_config_settings.R
\docType{data}
\name{output_options}
\alias{output_options}
\title{output_options}
\format{
Named \code{list} of output options.
}
\usage{
output_options
}
\description{
List of losstypes and variables allowed.
}
\keyword{datasets}
