% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputplots_module.R
\name{outputplotsUI}
\alias{outputplotsUI}
\alias{outputplots}
\title{outputplotsUI}
\usage{
outputplotsUI(id)

outputplots(
  input,
  output,
  session,
  selectPortfID,
  selectAnaID,
  n_panels,
  filesListData = reactive(NULL),
  active
)
}
\arguments{
\item{id}{Namespace ID for the module.}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{selectPortfID}{ID of selected portfolio.}

\item{selectAnaID}{ID of selected analysis.}

\item{n_panels}{Number of panels.}

\item{filesListData}{Table of output files for a given analysis ID.}

\item{active}{reactive expression whether the module state should be updated.}
}
\value{
List of tags.
}
\description{
UI/View for output plots of a run.

Server logic for outputplots of a run.
}
