% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pageheader_module.R
\name{pageheaderUI}
\alias{pageheaderUI}
\alias{pageheader}
\title{pageheaderUI}
\usage{
pageheaderUI(id)

pageheader(input, output, session, user, active = reactive(TRUE))
}
\arguments{
\item{id}{Namespace ID for the module.}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{user}{reactive expression yielding user name}

\item{active}{reactive expression whether the module state should be updated.}
}
\value{
List of tags.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

For \code{pageheader()}, list of reactives:
\itemize{
		\item{\code{logout}: }{reactive yielding logout button signal}
}.
}
\description{
pageheaderUI

pageheader
}
