% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_chooseAnalysis_module.R
\name{panelAnalysisDetails}
\alias{panelAnalysisDetails}
\title{panelAnalysisDetails}
\usage{
panelAnalysisDetails(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to show analyses details table.
}
