% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_configureOutput_module.R
\name{panelDefineOutputs}
\alias{panelDefineOutputs}
\title{panelDefineOutputs}
\usage{
panelDefineOutputs(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to define outputs
}
