% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputplots_module.R
\name{panelOutputModuleUI}
\alias{panelOutputModuleUI}
\alias{panelOutputModule}
\title{panelOutputModuleUI}
\usage{
panelOutputModuleUI(id)

panelOutputModule(
  input,
  output,
  session,
  portfId,
  anaID,
  filesListData = reactive(NULL),
  active
)
}
\arguments{
\item{id}{Namespace ID for the module.}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{portfId}{Portfolio ID.}

\item{anaID}{Analysis ID.}

\item{filesListData}{Table of output files for a given anaID.}

\item{active}{reactive expression whether the module state should be updated.}
}
\value{
Reactive value of the title.
}
\description{
Server logic to show graphical output such as plots.
}
