% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_config_module.R
\name{panelOutputParams}
\alias{panelOutputParams}
\title{panelOutputParams}
\usage{
panelOutputParams(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping sub-panel to define output params.
}
