% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_config_module.R
\name{panelSettingsTemplates}
\alias{panelSettingsTemplates}
\title{panelSettingsTemplates}
\usage{
panelSettingsTemplates(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping sub-panel to select previous analysis settings.
}
