% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_config_settings.R
\docType{data}
\name{plottypeslist}
\alias{plottypeslist}
\title{plottypeslist}
\format{
Named \code{list} of output plot options.
}
\usage{
plottypeslist
}
\description{
List containing settings for plots.
}
\keyword{datasets}
