% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasisuiPanel.R
\name{renderOasisuiPanelHeading}
\alias{renderOasisuiPanelHeading}
\title{Render Heading UI}
\usage{
renderOasisuiPanelHeading(expr, env = parent.frame(), ...)
}
\arguments{
\item{expr}{An expression that returns a Shiny tag object, \code{\link{HTML}},
or a list of such objects.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{...}{Further arguments passed to \code{\link[shiny:renderUI]{shiny::renderUI()}}.}
}
\description{
Render reactive panel heading content dynamically based on server logic.
}
\seealso{
\code{\link[=oasisuiPanelHeadingOutput]{oasisuiPanelHeadingOutput()}}
}
