% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_config_settings.R
\name{reportToVar}
\alias{reportToVar}
\title{reportToVar}
\usage{
reportToVar()
}
\description{
Function to create a named list connecting the file names patterns with the corresponding variables.
Operates on \code{varsdf}.
}
