% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputConfigUI_funs.R
\name{rerunUI}
\alias{rerunUI}
\title{Rerun UI}
\usage{
rerunUI(session, analysisID, tag, oed_field)
}
\arguments{
\item{session}{Current session.}

\item{analysisID}{Current analysis ID.}

\item{tag}{Current tag.}

\item{oed_field}{oed fields.}
}
\value{
UI function.
}
\description{
Summary Level and Reports fields in re-run situation.
}
