% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hub_helper_funcs.R
\name{simplify_path}
\alias{simplify_path}
\title{simplify_path}
\usage{
simplify_path(x, to_strip = NULL)
}
\arguments{
\item{x}{paths.}

\item{to_strip}{character vector of initial path to strip.}
}
\value{
simplified vector of paths.
}
\description{
Returns simplified paths.
}
\details{
directory paths ending with "/" are removed from set.
}
