% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleAnaWorkflowSteps.R
\name{singleAnaWorkflowStepsUI}
\alias{singleAnaWorkflowStepsUI}
\alias{singleAnaWorkflowSteps}
\title{singleAnaWorkflowStepsUI}
\usage{
singleAnaWorkflowStepsUI(id)

singleAnaWorkflowSteps(input, output, session)
}
\arguments{
\item{id}{Namespace ID for the module.}

\item{input, output, session}{Shiny input, output, and session objects.}
}
\value{
List of tags.

smth.
}
\description{
singleAnaWorkflowStepsUI

singleAnaWorkflowSteps
}
