% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_settings.R
\docType{data}
\name{Status_details}
\alias{Status_details}
\title{Status Details}
\format{
Named \code{list} of Staus details.
}
\usage{
Status_details
}
\description{
List of text representing status as returned by API
}
\keyword{datasets}
