% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_choosePortfolio_module.R
\name{step1_choosePortfolioUI}
\alias{step1_choosePortfolioUI}
\alias{step1_choosePortfolio}
\title{step1_choosePortfolioUI}
\usage{
step1_choosePortfolioUI(id)

step1_choosePortfolio(input, output, session, active = reactive(TRUE),
  currstep = reactive(-1), portfolioID = reactive(""))
}
\arguments{
\item{id}{module id}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{active}{reactive expression whether the module state should be updated.}

\item{currstep}{current selected step.}

\item{portfolioID}{selected portfolio ID.}

\item{id}{Namespace ID for the module.}
}
\value{
List of tags.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

portfolioID Id of selected portfolioID

tbl_portfoliosData POData model association table.

newstep navigation step
}
\description{
UI/View for the step1_choosePortfolio.

Server logic to step1_choosePortfolio
}
