% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_chooseAnalysis_module.R
\name{step2_chooseAnalysisUI}
\alias{step2_chooseAnalysisUI}
\alias{step2_chooseAnalysis}
\title{step2_chooseAnalysis UI}
\usage{
step2_chooseAnalysisUI(id)

step2_chooseAnalysis(input, output, session, active = reactive(TRUE),
  currstep = reactive(-1), portfolioID = reactive({     "" }),
  pfName = reactive({     "" }), pfstatus = reactive({     "" }))
}
\arguments{
\item{id}{module id}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{active}{reactive expression whether the module state should be updated.}

\item{currstep}{current selected step.}

\item{portfolioID}{selected portfolio ID.}

\item{pfName}{Name of selected portfolio}

\item{pfstatus}{Status of selected portfolio}

\item{id}{Namespace ID for the module.}
}
\value{
List of tags.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

newstep navigation step

analysisID analysis ID
}
\description{
UI/View for the step2_chooseAnalysis.

Server logic to step2_chooseAnalysis.
}
