% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_configureOutput_module.R
\name{step3_configureOutputUI}
\alias{step3_configureOutputUI}
\alias{step3_configureOutput}
\title{step3_configureOutputUI}
\usage{
step3_configureOutputUI(id)

step3_configureOutput(
  input,
  output,
  session,
  active = reactive(TRUE),
  currstep = reactive(-1),
  portfolioID = reactive(""),
  pfName = reactive(""),
  analysisID = reactive(NULL),
  customModSettings = reactive(NULL)
)
}
\arguments{
\item{id}{Namespace ID for the module.}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{active}{reactive expression whether the module state should be updated.}

\item{currstep}{Current selected step.}

\item{portfolioID}{Selected portfolio ID.}

\item{pfName}{Name of selected portfolio.}

\item{analysisID}{Selected analysis ID.}

\item{customModSettings}{Customizable Model settings.}
}
\value{
List of tags.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

dashboardAnaID id of selected run.
}
\description{
UI/View for the step3_configureOutput.

Server logic to step3_configureOutput.
}
