% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_config_settings.R
\docType{data}
\name{summary_template}
\alias{summary_template}
\title{summary_template}
\format{
Named \code{list} of output options.
}
\usage{
summary_template
}
\description{
List defining the structure of a "summary output" in the analysis settings.
}
\keyword{datasets}
