% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasisui_utils.R
\name{SwapNamesValueInList}
\alias{SwapNamesValueInList}
\title{Swap names and values of a list}
\usage{
SwapNamesValueInList(List)
}
\arguments{
\item{List}{list}
}
\value{
list
}
\description{
Returns a list with names and values switched with respect to input list.
}
