% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_settings.R
\docType{data}
\name{tbl_modelsDataNames}
\alias{tbl_modelsDataNames}
\title{tbl_modelsDataNames}
\format{
Named \code{list} of tbl_modelsData column names.
}
\usage{
tbl_modelsDataNames
}
\description{
Creating List for col names of Model Table
}
\keyword{datasets}
