% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_settings.R
\docType{data}
\name{tbl_portfoliosDataNames}
\alias{tbl_portfoliosDataNames}
\title{tbl_portfoliosDataNames}
\format{Named \code{list} of bl_portfoliosData column names.}
\usage{
tbl_portfoliosDataNames
}
\description{
Creating List for col names of Portfolio Table
}
\keyword{datasets}
