% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hub_helper_funcs.R
\name{untar_list}
\alias{untar_list}
\title{untar list}
\usage{
untar_list(tarfile, to_strip = NULL)
}
\arguments{
\item{tarfile}{path to tar bundle.}

\item{to_strip}{character vector of initial path to strip relative to the tar bundle content.}
}
\value{
list of files in tar bundle.
}
\description{
Returns list of files in a tar bundle.
}
