% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_config_settings.R
\docType{data}
\name{varsdf}
\alias{varsdf}
\title{varsdf}
\format{
\code{data.frame} of variables for output configuration.
}
\usage{
varsdf
}
\description{
Data frame of variables, with respective analysis_settings json field and info about whether a variable is under lec_output in analysis_settings.json.
}
\keyword{datasets}
