% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationBBR_server.R,
%   R/visualizationBBR_ui.R
\name{visualizationBBR}
\alias{visualizationBBR}
\alias{visualizationBBRUI}
\title{visualizationBBR}
\usage{
visualizationBBR(
  input,
  output,
  session,
  preselAnaId = reactive(NULL),
  anaID = reactive(NULL),
  active = reactive(TRUE)
)

visualizationBBRUI(id)
}
\arguments{
\item{input, output, session}{Shiny input, output, and session objects.}

\item{preselAnaId}{reactive string expression for reselected analysis id from \link{landingPage}.}

\item{anaID}{reactive string expression for reselected run id from \link{step3_configureOutput}.}

\item{active}{reactive expression whether the module state should be updated.}

\item{id}{Namespace ID for the module.}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

preselPanel panel to show in the model session

selectAnaID id of selected analysis

selectPortfolioID portfolio id of selected analysis

List of tags.
}
\description{
Server logic for batchbrowse run page.

UI/View for batchbrowse run page.
}
