% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationSBR_server.R,
%   R/visualizationSBR_ui.R
\name{visualizationSBR}
\alias{visualizationSBR}
\alias{visualizationSBRUI}
\title{visualizationSBR}
\usage{
visualizationSBR(input, output, session, preselAnaId = reactive(NULL),
  anaID = reactive(NULL), active = reactive(TRUE))

visualizationSBRUI(id)
}
\arguments{
\item{input, output, session}{Shiny input, output, and session objects.}

\item{preselAnaId}{reactive string expression for reselected analysis id from \link{landingPage}.}

\item{anaID}{reactive string expression for reselected run id from \link{step3_configureOutput}.}

\item{active}{reactive expression whether the module state should be updated.}

\item{id}{Namespace ID for the module.}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

preselPanel panel to show in the model session

selectAnaID id of selected analysis

selectPortfolioID portfolio id of selected analysis

List of tags.
}
\description{
Server logic for viewing results of a single analysis.

UI/View for viewing results of a single run.
}
