% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withModalSpinner.R
\name{withModalSpinner}
\alias{withModalSpinner}
\title{withModalSpinner}
\usage{
withModalSpinner(expr, info, spinner = icon("spinner", "fa-spin"),
  size = "m")
}
\arguments{
\item{expr}{The \code{expression} to be evaluated.}

\item{info}{The information message to be displayed.}

\item{spinner}{The spinning wheel icon.}

\item{size}{One of \code{"s"} for small, \code{"m"} (the default) for medium,
or \code{"l"} for large.}
}
\description{
Display a modal window with a spinning wheel and an information message
while a (time-consuming) expression is evaluated.
}
