% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hub_helper_funcs.R
\name{writefile}
\alias{writefile}
\title{write file}
\usage{
writefile(data, dataset_identifier = NULL, destdir = tempdir(),
  file_towrite = NULL)
}
\arguments{
\item{data}{object to write.}

\item{dataset_identifier}{name and relative path of file to write}

\item{destdir}{path where to write file.}

\item{file_towrite}{name of file where to wirte data.}
}
\description{
Writes ojbect in the correct format.
}
