% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelParams_funs.R
\name{advancedConfig_funs}
\alias{advancedConfig_funs}
\title{Advanced Output Configuration functions}
\usage{
advancedConfig_funs(session, model_settings)
}
\arguments{
\item{session}{Current session.}

\item{model_settings}{Model settings retireved from the API.}
}
\value{
List of functions.
}
\description{
Functions for the Advanced Output Configuration/Rerun Model Parameters view.
}
