% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelParams_funs.R
\name{basicConfig_funs}
\alias{basicConfig_funs}
\title{Basic Output Configuration functions}
\usage{
basicConfig_funs(session, model_settings)
}
\arguments{
\item{session}{Current session.}

\item{model_settings}{Model settings retireved from the API.}
}
\value{
List of functions.
}
\description{
Functions for the Basic Output Configuration/Rerun Model Parameters view.
}
