% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCustom_module.R
\name{buildCustom}
\alias{buildCustom}
\title{Hazard Map Module}
\usage{
buildCustom(input, output, session, portfolioID, modelID, supplierID,
  versionID, analysisID, counter, active = reactive(TRUE))
}
\arguments{
\item{input, output, session}{Shiny input, output, and session objects.}

\item{portfolioID}{Selected portfolio ID.}

\item{modelID}{Selected model ID.}

\item{supplierID}{Selected supplier model ID.}

\item{versionID}{Selected model ID version.}

\item{analysisID}{Selected analysis ID.}

\item{id}{Namespace ID for the module.}
}
\description{
Shiny Module for showing customized details.
}
\section{Functions}{
\itemize{
\item \code{buildCustom}: Allows user to build their own mode on the fly.
}}

