% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahub.R
\docType{class}
\name{DataHub}
\alias{DataHub}
\title{DataHub}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} with methods for Oasis files lists and Datatasets explorations.
}
\description{
R6 Class for OasisUI Data Hub.
}
\section{Arguments}{

\describe{
\item{id}{Portfolio Id | Model Id | Analysis Id}
\item{type}{Identifier for Analysis Id: inputs | outputs}
\item{dataset_identifier}{Identifier:  file name for Analysis Id |  api call for Portfolio Id and Model Id}
}
}

\section{Methods}{

\describe{

GENERAL

\item{\code{get_user_destdir()}}{return path of user specific foder}
\item{\code{oed_peril_codes()}}{return mapping of OED Peril names and codes.}

LISTS

> Portfolio
\item{\code{get_pf_data_list(id)}}{return list of portfolio source files}
\item{\code{invalidate_pf_data_list(id)}}{invalidate list of portfolio source files}
> Models
\item{\code{get_model_data_list(id)}}{return lists of model resources}
\item{\code{get_model_hazard_data_list(id)}}{return lists of model hazard resources}
> Analysis
\item{\code{get_ana_inputs_data_list(id)}}{return list of analysis inputs}
\item{\code{get_ana_outputs_data_list(id)}}{return list of analysis outputs}
\item{\code{invalidate_ana_inputs_data_list(id)}}{invalidate list of analysis inputs}
\item{\code{invalidate_ana_outputs_data_list(id)}}{invalidate list of analysis outputs}

DATASETS

> Portfolio
\item{\code{get_pf_dataset_content(id, dataset_identifier)}}{extract a source file (location/account...) content given a portfolio id}
\item{\code{invalidate_pf_dataset_content(id, dataset_identifier)}}{invalidate a source file (location/account...) content given a portfolio id}
\item{\code{get_pf_location_content(id)}}{extract location source file content given a portfolio id}
\item{\code{invalidate_pf_location_content(id)}}{invalidate a source file (location/account...) content given a portfolio id}
\item{\code{get_pf_dataset_header(id, dataset_identifier)}}{extract a source file (location/account...) header given a portfolio id}
\item{\code{invalidate_pf_dataset_header(id, dataset_identifier)}}{invalidate a source file (location/account...) header given a portfolio id}
\item{\code{get_pf_dataset_nrow(id, dataset_identifier)}}{extract a source file (location/account...) nrow given a portfolio id}
\item{\code{invalidate_pf_dataset_nrow(id, dataset_identifier)}}{invalidate a source file (location/account...) header given a portfolio id}
> Model
\item{\code{invalidate_model_resource_dataset_content(id)}}{invalidate model resource file given model id}
\item{\code{get_model_hazard_dataset_content(id)}}{extract model hazard resource file content given file id}
\item{\code{invalidate_model_hazard_dataset_content(id)}}{invalidate model hazard resource file content given file id}
> Analysis
\item{\code{get_ana_dataset_content(id, dataset_identifier, type)}}{extract a input/output file content given an analysis id}
\item{\code{invalidate_ana_dataset_content(id, dataset_identifier, type)}}{invalidate a input/output file content given an analysis id}
\item{\code{get_ana_inputs_dataset_content(id, dataset_identifier)}}{extract a inputs file content given an analysis id}
\item{\code{invalidate_ana_inputs_dataset_content(id,  dataset_identifier)}}{invalidate a inputs file content given an analysis id}
\item{\code{get_ana_outputs_dataset_content(id, dataset_identifier)}}{extract a outputs file content given an analysis id}
\item{\code{invalidate_ana_outputs_dataset_content(id,  dataset_identifier)}}{invalidate a outputs file content given an analysis id}
\item{\code{get_ana_dataset_header(id, type, dataset_identifier)}}{extract a inputs/outputs file nrow given an analysis id}
\item{\code{invalidate_ana_dataset_header(id, type, dataset_identifier)}}{invalidate a inputs/outputs file header given an analysis id}
\item{\code{get_ana_dataset_nrow(id, type, dataset_identifier)}}{extract a inputs/outputs file nrow given an analysis id}
\item{\code{invalidate_ana_dataset_nrow(id, type, dataset_identifier)}}{invalidate a inputs/outputs file nrow given an analysis id}
\item{\code{get_ana_dataset_size(id, type, dataset_identifier)}}{extract a inputs/outputs file size given an analysis id}
\item{\code{invalidate_ana_dataset_size(id, type, dataset_identifier)}}{invalidate a inputs/outputs file size given an analysis id}
\item{\code{get_ana_oed_summary_levels(id)}}{return list of valid oed summary levels for a given analysis}
\item{\code{invalidate_ana_oed_summay_levels(id)}}{invalidate list of valid oed summary levels for a given analysis}
\item{\code{get_ana_settings_content(id)}}{extract analysis settings content}
\item{\code{invalidate_ana_settings_content(id)}}{invalidate analysis settings content}
\item{\code{get_ana_validation_summary_content(id)}}{extract analysis validation summary content}
\item{\code{invalidate_ana_validation_summary_content(id)}}{invalidate analysis validation summary content}
> Write file
\item{\code{write_file()}}{Write data into a file file_towrite}
> Helper methods ----
\item{\code{get_analyses_tar(id, label, destdir = tempdir())}}{Extract input/output tar.}
\item{\code{get_analyses_inputs_tar(id, destdir = tempdir())}}{Extract input tar.}
\item{\code{get_analyses_outputs_tar(id, destdir = tempdir())}}{Extract output tar.}
> Return manipulated tables methods ----
# Portfolios
\item{\code{return_tbl_portfolioDetails(id)}}{Return dataframe of portfolio details.}
\item{\code{return_tbl_portfoliosData (name = "", Status, tbl_portfoliosDataNames)}}{Return data frame of portfolio data.}
# Models
\item{\code{return_tbl_modelsData(supplier_id = "", tbl_modelsDataNames)}}{Return data frame of models.}
\item{\code{return_tbl_modelData(id)}}{Return data frame of model data.}
# Analyses
\item{\code{return_analyses_input_file_wicons_df(id, Status)}}{Return analysis table with icons.}
\item{\code{return_tbl_analysesData_nice(tbl_analysesData, admin_mode, Status, tbl_modelsDataNames, tbl_portfoliosDataNames, tbl_analysesDataNames)}}{Return analysis table prettified.}
\item{\code{return_tbl_analysesData(name = "", Status, tbl_analysesDataNames)}}{Return dataframe of analyses.}
}
}

\section{Usage}{

\preformatted{data_hub <- DataHub$new()
data_hub$get_pf_data_list(id)
}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataHub$new()}}
\item \href{#method-terminate}{\code{DataHub$terminate()}}
\item \href{#method-get_user_destdir}{\code{DataHub$get_user_destdir()}}
\item \href{#method-get_oed_peril_codes_mapping}{\code{DataHub$get_oed_peril_codes_mapping()}}
\item \href{#method-get_pf_data_list}{\code{DataHub$get_pf_data_list()}}
\item \href{#method-invalidate_pf_data_list}{\code{DataHub$invalidate_pf_data_list()}}
\item \href{#method-get_model_data_list}{\code{DataHub$get_model_data_list()}}
\item \href{#method-get_model_hazard_data_list}{\code{DataHub$get_model_hazard_data_list()}}
\item \href{#method-get_ana_inputs_data_list}{\code{DataHub$get_ana_inputs_data_list()}}
\item \href{#method-get_ana_outputs_data_list}{\code{DataHub$get_ana_outputs_data_list()}}
\item \href{#method-invalidate_ana_inputs_data_list}{\code{DataHub$invalidate_ana_inputs_data_list()}}
\item \href{#method-invalidate_ana_outputs_data_list}{\code{DataHub$invalidate_ana_outputs_data_list()}}
\item \href{#method-get_pf_dataset_content}{\code{DataHub$get_pf_dataset_content()}}
\item \href{#method-invalidate_pf_dataset_content}{\code{DataHub$invalidate_pf_dataset_content()}}
\item \href{#method-get_pf_location_content}{\code{DataHub$get_pf_location_content()}}
\item \href{#method-invalidate_pf_location_content}{\code{DataHub$invalidate_pf_location_content()}}
\item \href{#method-get_pf_dataset_header}{\code{DataHub$get_pf_dataset_header()}}
\item \href{#method-invalidate_pf_dataset_header}{\code{DataHub$invalidate_pf_dataset_header()}}
\item \href{#method-get_pf_dataset_nrow}{\code{DataHub$get_pf_dataset_nrow()}}
\item \href{#method-invalidate_pf_dataset_nrow}{\code{DataHub$invalidate_pf_dataset_nrow()}}
\item \href{#method-invalidate_model_resource_dataset_content}{\code{DataHub$invalidate_model_resource_dataset_content()}}
\item \href{#method-get_model_hazard_dataset_content}{\code{DataHub$get_model_hazard_dataset_content()}}
\item \href{#method-invalidate_model_hazard_dataset_content}{\code{DataHub$invalidate_model_hazard_dataset_content()}}
\item \href{#method-get_ana_dataset_content}{\code{DataHub$get_ana_dataset_content()}}
\item \href{#method-invalidate_ana_dataset_content}{\code{DataHub$invalidate_ana_dataset_content()}}
\item \href{#method-get_ana_inputs_dataset_content}{\code{DataHub$get_ana_inputs_dataset_content()}}
\item \href{#method-invalidate_ana_inputs_dataset_content}{\code{DataHub$invalidate_ana_inputs_dataset_content()}}
\item \href{#method-get_ana_outputs_dataset_content}{\code{DataHub$get_ana_outputs_dataset_content()}}
\item \href{#method-invalidate_ana_outputs_dataset_content}{\code{DataHub$invalidate_ana_outputs_dataset_content()}}
\item \href{#method-get_ana_dataset_header}{\code{DataHub$get_ana_dataset_header()}}
\item \href{#method-invalidate_ana_dataset_header}{\code{DataHub$invalidate_ana_dataset_header()}}
\item \href{#method-get_ana_dataset_nrow}{\code{DataHub$get_ana_dataset_nrow()}}
\item \href{#method-invalidate_ana_dataset_nrow}{\code{DataHub$invalidate_ana_dataset_nrow()}}
\item \href{#method-get_ana_dataset_size}{\code{DataHub$get_ana_dataset_size()}}
\item \href{#method-invalidate_ana_dataset_size}{\code{DataHub$invalidate_ana_dataset_size()}}
\item \href{#method-get_ana_oed_summary_levels}{\code{DataHub$get_ana_oed_summary_levels()}}
\item \href{#method-invalidate_ana_oed_summay_levels}{\code{DataHub$invalidate_ana_oed_summay_levels()}}
\item \href{#method-get_ana_settings_content}{\code{DataHub$get_ana_settings_content()}}
\item \href{#method-invalidate_ana_settings_content}{\code{DataHub$invalidate_ana_settings_content()}}
\item \href{#method-get_ana_validation_summary_content}{\code{DataHub$get_ana_validation_summary_content()}}
\item \href{#method-get_ana_errors_summary_content}{\code{DataHub$get_ana_errors_summary_content()}}
\item \href{#method-invalidate_ana_validation_summary_content}{\code{DataHub$invalidate_ana_validation_summary_content()}}
\item \href{#method-write_file}{\code{DataHub$write_file()}}
\item \href{#method-write_parquet_file}{\code{DataHub$write_parquet_file()}}
\item \href{#method-get_analyses_tar}{\code{DataHub$get_analyses_tar()}}
\item \href{#method-get_analyses_inputs_tar}{\code{DataHub$get_analyses_inputs_tar()}}
\item \href{#method-get_analyses_outputs_tar}{\code{DataHub$get_analyses_outputs_tar()}}
\item \href{#method-return_tbl_portfolioDetails}{\code{DataHub$return_tbl_portfolioDetails()}}
\item \href{#method-return_tbl_portfoliosData}{\code{DataHub$return_tbl_portfoliosData()}}
\item \href{#method-return_tbl_modelsData}{\code{DataHub$return_tbl_modelsData()}}
\item \href{#method-return_tbl_modelData}{\code{DataHub$return_tbl_modelData()}}
\item \href{#method-return_analyses_input_file_wicons_df}{\code{DataHub$return_analyses_input_file_wicons_df()}}
\item \href{#method-return_tbl_analysesData_nice}{\code{DataHub$return_tbl_analysesData_nice()}}
\item \href{#method-return_tbl_analysesData}{\code{DataHub$return_tbl_analysesData()}}
\item \href{#method-return_tbl_dataFiles}{\code{DataHub$return_tbl_dataFiles()}}
\item \href{#method-clone}{\code{DataHub$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$new(user, destdir = tempdir(), oasisapi)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-terminate"></a>}}
\if{latex}{\out{\hypertarget{method-terminate}{}}}
\subsection{Method \code{terminate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$terminate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_user_destdir"></a>}}
\if{latex}{\out{\hypertarget{method-get_user_destdir}{}}}
\subsection{Method \code{get_user_destdir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_user_destdir()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_oed_peril_codes_mapping"></a>}}
\if{latex}{\out{\hypertarget{method-get_oed_peril_codes_mapping}{}}}
\subsection{Method \code{get_oed_peril_codes_mapping()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_oed_peril_codes_mapping()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_pf_data_list"></a>}}
\if{latex}{\out{\hypertarget{method-get_pf_data_list}{}}}
\subsection{Method \code{get_pf_data_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_pf_data_list(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invalidate_pf_data_list"></a>}}
\if{latex}{\out{\hypertarget{method-invalidate_pf_data_list}{}}}
\subsection{Method \code{invalidate_pf_data_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$invalidate_pf_data_list(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_model_data_list"></a>}}
\if{latex}{\out{\hypertarget{method-get_model_data_list}{}}}
\subsection{Method \code{get_model_data_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_model_data_list(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_model_hazard_data_list"></a>}}
\if{latex}{\out{\hypertarget{method-get_model_hazard_data_list}{}}}
\subsection{Method \code{get_model_hazard_data_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_model_hazard_data_list(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ana_inputs_data_list"></a>}}
\if{latex}{\out{\hypertarget{method-get_ana_inputs_data_list}{}}}
\subsection{Method \code{get_ana_inputs_data_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_ana_inputs_data_list(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ana_outputs_data_list"></a>}}
\if{latex}{\out{\hypertarget{method-get_ana_outputs_data_list}{}}}
\subsection{Method \code{get_ana_outputs_data_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_ana_outputs_data_list(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invalidate_ana_inputs_data_list"></a>}}
\if{latex}{\out{\hypertarget{method-invalidate_ana_inputs_data_list}{}}}
\subsection{Method \code{invalidate_ana_inputs_data_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$invalidate_ana_inputs_data_list(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invalidate_ana_outputs_data_list"></a>}}
\if{latex}{\out{\hypertarget{method-invalidate_ana_outputs_data_list}{}}}
\subsection{Method \code{invalidate_ana_outputs_data_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$invalidate_ana_outputs_data_list(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_pf_dataset_content"></a>}}
\if{latex}{\out{\hypertarget{method-get_pf_dataset_content}{}}}
\subsection{Method \code{get_pf_dataset_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_pf_dataset_content(id, dataset_identifier, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invalidate_pf_dataset_content"></a>}}
\if{latex}{\out{\hypertarget{method-invalidate_pf_dataset_content}{}}}
\subsection{Method \code{invalidate_pf_dataset_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$invalidate_pf_dataset_content(id, dataset_identifier, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_pf_location_content"></a>}}
\if{latex}{\out{\hypertarget{method-get_pf_location_content}{}}}
\subsection{Method \code{get_pf_location_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_pf_location_content(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invalidate_pf_location_content"></a>}}
\if{latex}{\out{\hypertarget{method-invalidate_pf_location_content}{}}}
\subsection{Method \code{invalidate_pf_location_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$invalidate_pf_location_content(id, dataset_identifier, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_pf_dataset_header"></a>}}
\if{latex}{\out{\hypertarget{method-get_pf_dataset_header}{}}}
\subsection{Method \code{get_pf_dataset_header()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_pf_dataset_header(id, dataset_identifier, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invalidate_pf_dataset_header"></a>}}
\if{latex}{\out{\hypertarget{method-invalidate_pf_dataset_header}{}}}
\subsection{Method \code{invalidate_pf_dataset_header()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$invalidate_pf_dataset_header(id, dataset_identifier, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_pf_dataset_nrow"></a>}}
\if{latex}{\out{\hypertarget{method-get_pf_dataset_nrow}{}}}
\subsection{Method \code{get_pf_dataset_nrow()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_pf_dataset_nrow(id, dataset_identifier, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invalidate_pf_dataset_nrow"></a>}}
\if{latex}{\out{\hypertarget{method-invalidate_pf_dataset_nrow}{}}}
\subsection{Method \code{invalidate_pf_dataset_nrow()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$invalidate_pf_dataset_nrow(id, dataset_identifier, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invalidate_model_resource_dataset_content"></a>}}
\if{latex}{\out{\hypertarget{method-invalidate_model_resource_dataset_content}{}}}
\subsection{Method \code{invalidate_model_resource_dataset_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$invalidate_model_resource_dataset_content(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_model_hazard_dataset_content"></a>}}
\if{latex}{\out{\hypertarget{method-get_model_hazard_dataset_content}{}}}
\subsection{Method \code{get_model_hazard_dataset_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_model_hazard_dataset_content(id, filename, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invalidate_model_hazard_dataset_content"></a>}}
\if{latex}{\out{\hypertarget{method-invalidate_model_hazard_dataset_content}{}}}
\subsection{Method \code{invalidate_model_hazard_dataset_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$invalidate_model_hazard_dataset_content(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ana_dataset_content"></a>}}
\if{latex}{\out{\hypertarget{method-get_ana_dataset_content}{}}}
\subsection{Method \code{get_ana_dataset_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_ana_dataset_content(id, dataset_identifier, type = "", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invalidate_ana_dataset_content"></a>}}
\if{latex}{\out{\hypertarget{method-invalidate_ana_dataset_content}{}}}
\subsection{Method \code{invalidate_ana_dataset_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$invalidate_ana_dataset_content(id, dataset_identifier, type, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ana_inputs_dataset_content"></a>}}
\if{latex}{\out{\hypertarget{method-get_ana_inputs_dataset_content}{}}}
\subsection{Method \code{get_ana_inputs_dataset_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_ana_inputs_dataset_content(id, dataset_identifier, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invalidate_ana_inputs_dataset_content"></a>}}
\if{latex}{\out{\hypertarget{method-invalidate_ana_inputs_dataset_content}{}}}
\subsection{Method \code{invalidate_ana_inputs_dataset_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$invalidate_ana_inputs_dataset_content(id, dataset_identifier, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ana_outputs_dataset_content"></a>}}
\if{latex}{\out{\hypertarget{method-get_ana_outputs_dataset_content}{}}}
\subsection{Method \code{get_ana_outputs_dataset_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_ana_outputs_dataset_content(id, dataset_identifier, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invalidate_ana_outputs_dataset_content"></a>}}
\if{latex}{\out{\hypertarget{method-invalidate_ana_outputs_dataset_content}{}}}
\subsection{Method \code{invalidate_ana_outputs_dataset_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$invalidate_ana_outputs_dataset_content(id, dataset_identifier, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ana_dataset_header"></a>}}
\if{latex}{\out{\hypertarget{method-get_ana_dataset_header}{}}}
\subsection{Method \code{get_ana_dataset_header()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_ana_dataset_header(id, type, dataset_identifier, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invalidate_ana_dataset_header"></a>}}
\if{latex}{\out{\hypertarget{method-invalidate_ana_dataset_header}{}}}
\subsection{Method \code{invalidate_ana_dataset_header()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$invalidate_ana_dataset_header(id, type, dataset_identifier, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ana_dataset_nrow"></a>}}
\if{latex}{\out{\hypertarget{method-get_ana_dataset_nrow}{}}}
\subsection{Method \code{get_ana_dataset_nrow()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_ana_dataset_nrow(id, type, dataset_identifier, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invalidate_ana_dataset_nrow"></a>}}
\if{latex}{\out{\hypertarget{method-invalidate_ana_dataset_nrow}{}}}
\subsection{Method \code{invalidate_ana_dataset_nrow()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$invalidate_ana_dataset_nrow(id, type, dataset_identifier, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ana_dataset_size"></a>}}
\if{latex}{\out{\hypertarget{method-get_ana_dataset_size}{}}}
\subsection{Method \code{get_ana_dataset_size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_ana_dataset_size(id, type, dataset_identifier, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invalidate_ana_dataset_size"></a>}}
\if{latex}{\out{\hypertarget{method-invalidate_ana_dataset_size}{}}}
\subsection{Method \code{invalidate_ana_dataset_size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$invalidate_ana_dataset_size(id, type, dataset_identifier, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ana_oed_summary_levels"></a>}}
\if{latex}{\out{\hypertarget{method-get_ana_oed_summary_levels}{}}}
\subsection{Method \code{get_ana_oed_summary_levels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_ana_oed_summary_levels(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invalidate_ana_oed_summay_levels"></a>}}
\if{latex}{\out{\hypertarget{method-invalidate_ana_oed_summay_levels}{}}}
\subsection{Method \code{invalidate_ana_oed_summay_levels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$invalidate_ana_oed_summay_levels(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ana_settings_content"></a>}}
\if{latex}{\out{\hypertarget{method-get_ana_settings_content}{}}}
\subsection{Method \code{get_ana_settings_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_ana_settings_content(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invalidate_ana_settings_content"></a>}}
\if{latex}{\out{\hypertarget{method-invalidate_ana_settings_content}{}}}
\subsection{Method \code{invalidate_ana_settings_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$invalidate_ana_settings_content(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ana_validation_summary_content"></a>}}
\if{latex}{\out{\hypertarget{method-get_ana_validation_summary_content}{}}}
\subsection{Method \code{get_ana_validation_summary_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_ana_validation_summary_content(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ana_errors_summary_content"></a>}}
\if{latex}{\out{\hypertarget{method-get_ana_errors_summary_content}{}}}
\subsection{Method \code{get_ana_errors_summary_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_ana_errors_summary_content(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invalidate_ana_validation_summary_content"></a>}}
\if{latex}{\out{\hypertarget{method-invalidate_ana_validation_summary_content}{}}}
\subsection{Method \code{invalidate_ana_validation_summary_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$invalidate_ana_validation_summary_content(
  id,
  dataset_identifier,
  type,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_file"></a>}}
\if{latex}{\out{\hypertarget{method-write_file}{}}}
\subsection{Method \code{write_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$write_file(data, dataset_identifier, file_towrite = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_parquet_file"></a>}}
\if{latex}{\out{\hypertarget{method-write_parquet_file}{}}}
\subsection{Method \code{write_parquet_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$write_parquet_file(data, dataset_identifier, file_towrite = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_analyses_tar"></a>}}
\if{latex}{\out{\hypertarget{method-get_analyses_tar}{}}}
\subsection{Method \code{get_analyses_tar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_analyses_tar(id, label, destdir = tempdir())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_analyses_inputs_tar"></a>}}
\if{latex}{\out{\hypertarget{method-get_analyses_inputs_tar}{}}}
\subsection{Method \code{get_analyses_inputs_tar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_analyses_inputs_tar(id, destdir = tempdir())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_analyses_outputs_tar"></a>}}
\if{latex}{\out{\hypertarget{method-get_analyses_outputs_tar}{}}}
\subsection{Method \code{get_analyses_outputs_tar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$get_analyses_outputs_tar(id, destdir = tempdir())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-return_tbl_portfolioDetails"></a>}}
\if{latex}{\out{\hypertarget{method-return_tbl_portfolioDetails}{}}}
\subsection{Method \code{return_tbl_portfolioDetails()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$return_tbl_portfolioDetails(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-return_tbl_portfoliosData"></a>}}
\if{latex}{\out{\hypertarget{method-return_tbl_portfoliosData}{}}}
\subsection{Method \code{return_tbl_portfoliosData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$return_tbl_portfoliosData(name = "", Status, tbl_portfoliosDataNames)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-return_tbl_modelsData"></a>}}
\if{latex}{\out{\hypertarget{method-return_tbl_modelsData}{}}}
\subsection{Method \code{return_tbl_modelsData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$return_tbl_modelsData(supplier_id = "", tbl_modelsDataNames)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-return_tbl_modelData"></a>}}
\if{latex}{\out{\hypertarget{method-return_tbl_modelData}{}}}
\subsection{Method \code{return_tbl_modelData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$return_tbl_modelData(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-return_analyses_input_file_wicons_df"></a>}}
\if{latex}{\out{\hypertarget{method-return_analyses_input_file_wicons_df}{}}}
\subsection{Method \code{return_analyses_input_file_wicons_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$return_analyses_input_file_wicons_df(id, Status)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-return_tbl_analysesData_nice"></a>}}
\if{latex}{\out{\hypertarget{method-return_tbl_analysesData_nice}{}}}
\subsection{Method \code{return_tbl_analysesData_nice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$return_tbl_analysesData_nice(
  tbl_analysesData,
  admin_mode,
  Status,
  tbl_modelsDataNames,
  tbl_portfoliosDataNames,
  tbl_analysesDataNames
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-return_tbl_analysesData"></a>}}
\if{latex}{\out{\hypertarget{method-return_tbl_analysesData}{}}}
\subsection{Method \code{return_tbl_analysesData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$return_tbl_analysesData(name = "", Status, tbl_analysesDataNames)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-return_tbl_dataFiles"></a>}}
\if{latex}{\out{\hypertarget{method-return_tbl_dataFiles}{}}}
\subsection{Method \code{return_tbl_dataFiles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$return_tbl_dataFiles(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHub$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
