% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputConfigUI_funs.R
\name{dynamicUI_btns}
\alias{dynamicUI_btns}
\title{Dynamic UI with buttons}
\usage{
dynamicUI_btns(session, analysisID, ana_flag, tag, oed_field)
}
\arguments{
\item{session}{Current session.}

\item{analysisID}{Current analysis ID.}

\item{ana_flag}{Analysis flag.}

\item{tag}{Current tag.}

\item{oed_field}{oed fields.}
}
\value{
UI function.
}
\description{
Add "+" and "x" buttons to dynamic UI.
}
