% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelParams_funs.R
\name{Global_funs}
\alias{Global_funs}
\title{Global Model Settings parameters function}
\usage{
Global_funs(session, model_settings, ui_step)
}
\arguments{
\item{session}{Current session.}

\item{model_settings}{Model settings retrieved from the API.}

\item{ui_step}{Step 2}
}
\value{
List of UI elements (input/selector widgets).
}
\description{
Functions for the Build Custom UI.
}
