% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_text.R
\docType{data}
\name{helper-text}
\alias{helper-text}
\alias{file_Viewer_tooltips}
\alias{landing_page_tooltips}
\alias{defineSingleAna_tooltips}
\alias{dashboard_tooltips}
\title{Helper text}
\format{
Each object is a named list of character strings, the names
  corresponding to the button IDs.

An object of class \code{list} of length 1.

An object of class \code{list} of length 13.

An object of class \code{list} of length 26.

An object of class \code{list} of length 2.
}
\usage{
file_Viewer_tooltips

landing_page_tooltips

defineSingleAna_tooltips

dashboard_tooltips
}
\description{
Helper texts to show explanations next to buttons. Grouped according to where the button is visible in the UI.
}
\keyword{datasets}
