% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{oasisuiTableButton}
\alias{oasisuiTableButton}
\title{oasisuiTableButton}
\usage{
oasisuiTableButton(
  inputId,
  label,
  icon = NULL,
  width = NULL,
  class = c("btn", "btn-secondary"),
  ...
)
}
\arguments{
\item{inputId}{widget id}

\item{label}{widget label}

\item{icon}{widget icon}

\item{width}{widget width}

\item{class}{HTML class attribute.}

\item{...}{Arguments to \code{\link[shiny:actionButton]{shiny::actionButton()}}.}
}
\value{
Button widget (UI element).
}
\description{
Modified version of the primary \code{\link[=oasisuiButton]{oasisuiButton()}}.
}
