% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_config_settings.R
\name{reportToVar}
\alias{reportToVar}
\title{reportToVar}
\usage{
reportToVar(varsdf)
}
\arguments{
\item{varsdf}{dataframe of variables for output configuration. Defined in the package.}
}
\description{
function to create a nemed list connecting the file names patterns with the coorresponding vaariables. takes as input varsdf
}
