% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputConfigUI_funs.R
\name{rerunUI}
\alias{rerunUI}
\title{Rerun UI}
\usage{
rerunUI(session, analysisID, tag, oed_field, analysisSettings = NULL)
}
\arguments{
\item{session}{Current session.}

\item{analysisID}{Current analysis ID.}

\item{tag}{Current tag.}

\item{oed_field}{Oed fields.}

\item{analysisSettings}{Analysis settings for case where we re-run from settings template.}
}
\value{
UI element.
}
\description{
Summary level and reports fields in re-run situation.
}
