% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_configureOutput_module.R
\name{step3_configureOutputUI}
\alias{step3_configureOutputUI}
\alias{step3_configureOutput}
\title{step3_configureOutputUI}
\usage{
step3_configureOutputUI(id)

step3_configureOutput(input, output, session, active = reactive(TRUE),
  currstep = reactive(-1), portfolioID = reactive(""),
  pfName = reactive(""), analysisID = reactive(NULL))
}
\arguments{
\item{id}{Namespace ID for the module.}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{active}{reactive expression whether the module state should be updated.}

\item{currstep}{Current selected step.}

\item{portfolioID}{Selected portfolio ID.}

\item{pfName}{Name of selected portfolio.}

\item{analysisID}{Selected analysis ID.}
}
\value{
List of tags.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

dashboardAnaID id of selected run.
}
\description{
UI/View for the step3_configureOutput.

Server logic to step3_configureOutput.
}
\details{
#' panelDefineOutputsDetails
#'
#' @rdname panelDefineOutputsDetails
#'
#' @description Function wrapping sub-panel to define outputs details.
#'
#' @template params-module-ui
#'
#' @importFrom shinyjs hidden
#'
#' @export
panelDefineOutputsDetails <- function(id) {
  ns <- NS(id)
  tagList(
    oasisuiPanel(
      collapsible = FALSE,
      ns("panel_ConfigDetails"),
      heading = h4("Configuration details"),
      selectInput(ns("sinoutputoptions"), "Select Custom Configuration:", choices = "")
    ),
    oasisuiPanel(
      collapsible = FALSE,
      ns("panel_defAnaOutputDetails"),
      heading = h4("Model parameters"),
      div(id = ns("basic"), style = "width:100%; margin: 0 auto;",
          uiOutput(ns("basic_model_param")),
          uiOutput(ns("chkinputsperils"))
      ),
      hidden(div(id = ns("configureAnaParamsAdvanced"), align = "left",
                 textInput(ns("tinputnoofsample"), label = "Number of Samples:", value = "10"),
                 textInput(ns("tinputthreshold"), label = "Loss Threshold:", value = "0"),
                 checkboxInput(ns("chkinputsummaryoption"), "Summary Reports", value = TRUE),
                 uiOutput(ns("advanced_model_param"))
      ))
    )
  )
}

#' panelDefOutputConfiguration
#'
#' @rdname panelDefOutputConfiguration
#'
#' @description Function wrapping sub-panel to define outputs configuration.
#'
#' @template params-module-ui
#'
#' @importFrom shinyjs hidden
#' @importFrom bsplus bs_embed_tooltip
#'
#' @export
panelDefOutputConfiguration <- function(id) {
  ns <- NS(id)
  oasisuiPanel(
    collapsible = FALSE,
    ns("panel_outconfig"),
    heading = h4("Output configuration"),
    checkboxInput(ns("chkinputGUL"), label = "Ground Up Loss", value = TRUE),
    hidden(div(id = ns("panel_configureAdvancedGUL"), panel_configureAdvancedGUL(id))),
    checkboxInput(ns("chkinputIL"), label = "Insured Loss", value = FALSE),
    hidden(div(id = ns("panel_configureAdvancedIL"), panel_configureAdvancedIL(id))),
    checkboxInput(ns("chkinputRI"), label = "Net RI Loss", value = FALSE),
    hidden(div(id = ns("panel_configureAdvancedRI"), panel_configureAdvancedRI(id))),
    oasisuiButton(inputId = ns("abuttonadvanced"), label = "Advanced"),
    hidden(oasisuiButton(inputId = ns("abuttonbasic"), label = "Basic")),
    hidden(oasisuiButton(inputId = ns("abuttonclroutopt"), label = "Default"))
  )
}

#' panel_configureAdvancedGUL
#'
#' @rdname panel_configureAdvancedGUL
#'
#' @description Function wrapping sub-panel to define outputs advanced configuration GUL.
#'
#' @template params-module-ui
#'
#' @export
panel_configureAdvancedGUL <- function(id) {
  ns <- NS(id)

fluidRow(
    # Few outputs commented/disabled for the first release. To be enabled for later releases.
    column(4,
           h4("Ground Up Loss", class = "oasisui-loss"),
           h5("Full Sample", class = "oasisui-measure"),
           h5("ELT", class = "oasisui-measure"),
           tags$div(class = "h5-align", h5("AEP", class = "oasisui-measure")),
           tags$div(class = "h5-align", h5("OEP", class = "oasisui-measure")),
           tags$div(class = "h5-align", h5("Multi AEP", class = "oasisui-measure")),
           h5("Multi OEP", class = "oasisui-measure"),
           # h5("WS Mean AEP", class = "oasisui-measure"),
           # tags$div(class = "h5-align", h5("WS Mean OEP", class = "oasisui-measure")),
           # tags$div(class = "h5-align", h5("Sample Mean AEP", class = "oasisui-measure")),
           # h5("Sample Mean OEP", class = "oasisui-measure"),
           h5("AAL", class = "oasisui-measure"),
           tags$div(class = "h5-align", h5("PLT", class = "oasisui-measure"))),

tags$div(class = "multicol",
             checkboxGroupInput(ns("chkgulprog"),
                                label = h6("Prog", class = "oasisui-granularity"),
                                choices = LosstypesChoices,
                                selected = varsdf$vars[varsdf$defaultChoice]),

checkboxGroupInput(ns("chkgulstate"),
                                label = h6("State", class = "oasisui-granularity"),
                                choices = LosstypesChoices,
                                selected = NULL),

checkboxGroupInput(ns("chkgulcounty"),
                                label = h6("County", class = "oasisui-granularity"),
                                choices =LosstypesChoices,
                                selected = NULL),

checkboxGroupInput(ns("chkgulloc"),
                                label = h6("Location", class = "oasisui-granularity"),
                                choices = LosstypesChoices,
                                selected = NULL),

panelDefOutputConfiguration

checkboxGroupInput(ns("chkgulpolicy"),
                                label = h6("Policy", class = "oasisui-granularity"),
                                choices = LosstypesChoices,
                                selected = NULL)
    )
  )
}

#' panel_configureAdvancedIL
#'
#' @rdname panel_configureAdvancedIL
#'
#' @description Function wrapping sub-panel to define outputs advanced configuration IL.
#'
#' @template params-module-ui
#'
#' @export
panel_configureAdvancedIL <- function(id) {
  ns <- NS(id)
  fluidRow(
    column(4,
           h4("Insured Loss", class = "oasisui-loss"),
           h5("Full Sample", class = "oasisui-measure"),
           h5("ELT", class = "oasisui-measure"),
           tags$div(class = "h5-align", h5("AEP", class = "oasisui-measure")),
           tags$div(class = "h5-align", h5("OEP", class = "oasisui-measure")),
           tags$div(class = "h5-align", h5("Multi AEP", class = "oasisui-measure")),
           h5("Multi OEP", class = "oasisui-measure"),
           # h5("WS Mean AEP", class = "oasisui-measure"),
           # tags$div(class = "h5-align", h5("WS Mean OEP", class = "oasisui-measure")),
           # tags$div(class = "h5-align", h5("Sample Mean AEP", class = "oasisui-measure")),
           # h5("Sample Mean OEP", class = "oasisui-measure"),
           h5("AAL", class = "oasisui-measure"),
           tags$div(class = "h5-align", h5("PLT", class = "oasisui-measure"))),

tags$div(class = "multicol",
             checkboxGroupInput(ns("chkilprog"),
                                label = h6("Prog", class = "oasisui-granularity"),
                                choices = LosstypesChoices,
                                selected = NULL),

checkboxGroupInput(ns("chkilstate"),
                                label = h6("State", class = "oasisui-granularity"),
                                choices = LosstypesChoices,
                                selected = NULL),

checkboxGroupInput(ns("chkilcounty"),
                                label = h6("County", class = "oasisui-granularity"),
                                choices = LosstypesChoices,
                                selected = NULL),

checkboxGroupInput(ns("chkilloc"),
                                label = h6("Location", class = "oasisui-granularity"),
                                choices = LosstypesChoices,
                                selected = NULL),

checkboxGroupInput(ns("chkillob"),
                                label = h6("LOB", class = "oasisui-granularity"),
                                choices = LosstypesChoices,
                                selected = NULL),

checkboxGroupInput(ns("chkilpolicy"),
                                label = h6("Policy", class = "oasisui-granularity"),
                                choices = LosstypesChoices,
                                selected = NULL)
    )
  )
}

#' panel_configureAdvancedRI
#'
#' @rdname panel_configureAdvancedRI
#'
#' @description Function wrapping sub-panel to define outputs advanced configuration RI.
#'
#' @template params-module-ui
#'
#' @export
panel_configureAdvancedRI <- function(id) {
  ns <- NS(id)
  fluidRow(
    column(4,
           h4("Net RI Loss", style = "font-size: 18px; font-weight: bold;"),
           h5("Full Sample", style = "font-size: 16.5px;"),
           h5("ELT", style="font-size: 16.5px;"),
           tags$div(class = "h5-align", h5("AEP", style="font-size: 16.5px;")),
           tags$div(class = "h5-align", h5("OEP", style="font-size: 16.5px;")),
           tags$div(class = "h5-align", h5("Multi AEP", style="font-size: 16.5px;")),
           h5("Multi OEP", style = "font-size: 16.5px;"),
           # h5("WS Mean AEP", style="font-size: 16.5px;"),
           # tags$div(class = "h5-align", h5("WS Mean OEP", style="font-size: 16.5px;")),
           # tags$div(class = "h5-align",h5("Sample Mean AEP", style="font-size: 16.5px;")),
           # h5("Sample Mean OEP", style="font-size: 16.5px;"),
           h5("AAL", style="font-size: 16.5px;"),
           tags$div(class = "h5-align",h5("PLT", style="font-size: 16.5px;"))),

tags$div(class = "multicol",
             checkboxGroupInput(ns("chkriprog"),
                                label = h5("Prog", style = "font-size: 15.0px;"),
                                choices = LosstypesChoices,
                                selected = NULL),

checkboxGroupInput(ns("chkristate"),
                                label = h5("State", style = "font-size: 15.0px;"),
                                choices = LosstypesChoices,
                                selected = NULL),

checkboxGroupInput(ns("chkricounty"),
                                label = h5("County", style = "font-size: 15.0px;"),
                                choices = LosstypesChoices,
                                selected = NULL),

checkboxGroupInput(ns("chkriloc"),
                                label = h5("Location", style = "font-size: 15.0px;"),
                                choices = LosstypesChoices,
                                selected = NULL),

checkboxGroupInput(ns("chkrilob"),
                                label = h5("LOB", style = "font-size: 15.0px;"),
                                choices = LosstypesChoices,
                                selected = NULL),

checkboxGroupInput(ns("chkripolicy"),
                                label = h5("Policy", style = "font-size: 15.0px;"),
                                choices = LosstypesChoices,
                                selected = NULL)
    )
  )#end of fluidrow RI
}

step3_configureOutput server
}
